\docType{data}
\name{bb}
\alias{bb}
\alias{bb1112}
\title{The Big Bang Theory}
\value{
  The data set is a data frame

  The data set is a data frame
}
\description{
  The results from guessing the age of actors and actresses
  from The Bing Theory. The worksheet and slides from the
  lecture can be found at
  \url{http://www.mas.ncl.ac.uk/~ncsg3/teaching/mas1343/bb.pdf}
  and
  \url{http://www.mas.ncl.ac.uk/~ncsg3/teaching/mas1343/bb_ages.pdf}

  The results from guessing the age of actors and actresses
  from The Bing Theory. The worksheet and slides from the
  lecture can be found at
  \url{http://www.mas.ncl.ac.uk/~ncsg3/teaching/mas1343/bb.pdf}
  and
  \url{http://www.mas.ncl.ac.uk/~ncsg3/teaching/mas1343/bb_ages.pdf}
}
\note{
  The data set is a data frame with the following columns:
  \describe{ \item{ID}{A unique group id} \item{Name}{Team
  name} \item{Amy}{Guessed age for Amy}
  \item{Howard}{Guessed age for Howard}
  \item{Leonard}{Guessed age for Leonard}
  \item{Penny}{Guessed age for Penny} \item{Raj}{Guessed
  age for Raj} \item{Shel}{Guessed age for Sheldon}
  \item{Watch}{Has anyone in the team watch The Big Bang
  Theory before} \item{Gender}{The gender of the majority
  of the team} }

  The data set is a data frame with the following columns:
  \describe{ \item{ID}{A unique group id} \item{Name}{Team
  name} \item{Amy}{Guessed age for Amy}
  \item{Howard}{Guessed age for Howard}
  \item{Leonard}{Guessed age for Leonard}
  \item{Penny}{Guessed age for Penny} \item{Raj}{Guessed
  age for Raj} \item{Shel}{Guessed age for Sheldon}
  \item{Watch}{Has anyone in the team watch The Big Bang
  Theory before} \item{Gender}{The gender of the majority
  of the team} }
}
\keyword{datasets}

