\name{GetNumericVector}
\alias{GetCharacterVector}
\alias{GetLogicalVector}
\alias{GetNumericVector}
\title{Generate random vectors}
\usage{
  GetNumericVector(student_id)

  GetLogicalVector(student_id)

  GetCharacterVector(student_id)
}
\arguments{
  \item{student_id,}{a student id, i.e. b1234567}
}
\description{
  These functions create vectors of numerics, logicals or
  characters. Given a student id, the functions will always
  return the same subset.
}

