\docType{package}
\name{mas1343-package}
\alias{mas1343}
\alias{MAS1343}
\alias{mas1343-package}
\title{MAS1343}
\description{
  MAS1343
}
\details{
  Functions and datasets used in MAS1343: Introduction to
  statistical computing with R
}
\author{
  \email{colin.gillespie@newcastle.ac.uk}
}
\seealso{
  See the Blackboard homepage for further details.
}
\keyword{package}

