\name{CheckFunctionName}
\alias{CheckFunctionName}
\title{Checks file name}
\usage{
  CheckFunctionName(path)
}
\arguments{
  \item{path}{the path to where the file is located on your
  computer}
}
\value{
  A message describing whether the file has been named
  correctly.
}
\description{
  This function is used in practical 3 and checks that the
  file has been name correctly.
}
\examples{
\dontrun{
CheckFunctionName("practical3.r")
}
}

